linux-gen: pktio: move ethtool rss function to a new file
Cleanup socket pktio code by moving independent rss functionality
to a separate file.
