+++ platform/linux-generic/Makefile.am
@@ -113,6 +113,7 @@ noinst_HEADERS = \
 		  include/odp_packet_io_ring_internal.h \
 		  include/odp_packet_socket.h \
 		  include/odp_ethtool_stats.h \
+		  include/odp_ethtool_rss.h \
 		  include/odp_pcapng.h \
 		  include/odp_pkt_queue_internal.h \
 		  include/odp_pool_internal.h \
@@ -201,6 +202,7 @@ __LIB__libodp_linux_la_SOURCES = \
 			   pktio/dpdk.c \
 			   pktio/dpdk_parse.c \
 			   pktio/ethtool_stats.c \
+			   pktio/ethtool_rss.c \
 			   pktio/io_ops.c \
 			   pktio/ipc.c \
 			   pktio/loop.c \
