+++ platform/linux-generic/pktio/netmap.c
@@ -17,6 +17,7 @@
 
 #include <odp_packet_io_internal.h>
 #include <odp_ethtool_stats.h>
+#include <odp_ethtool_rss.h>
 #include <odp_packet_socket.h>
 #include <odp_debug_internal.h>
 #include <odp_errno_define.h>
